Refactor FileCachePlugin into multiple modules
