+++ lib/WebpackOptionsApply.js
@@ -471,7 +471,8 @@ class WebpackOptionsApply extends OptionsApply {
 		new WarnCaseSensitiveModulesPlugin().apply(compiler);
 
 		if (options.cache && typeof options.cache === "object") {
-			switch (options.cache.type) {
+			const cacheOptions = options.cache;
+			switch (cacheOptions.type) {
 				case "memory": {
 					const MemoryCachePlugin = require("./cache/MemoryCachePlugin");
 					new MemoryCachePlugin().apply(compiler);
@@ -480,8 +481,59 @@ class WebpackOptionsApply extends OptionsApply {
 				case "filesystem": {
 					const MemoryCachePlugin = require("./cache/MemoryCachePlugin");
 					new MemoryCachePlugin().apply(compiler);
-					const FileCachePlugin = require("./cache/FileCachePlugin");
-					new FileCachePlugin(options.cache).apply(compiler);
+					switch (cacheOptions.store) {
+						case "instant": {
+							const InstantFileCachePlugin = require("./cache/InstantFileCachePlugin");
+							const SeparateFilesCacheStrategy = require("./cache/SeparateFilesCacheStrategy");
+							new InstantFileCachePlugin(
+								new SeparateFilesCacheStrategy({
+									cacheLocation: cacheOptions.cacheLocation,
+									version: cacheOptions.version,
+									hashAlgorithm: cacheOptions.hashAlgorithm,
+									loglevel: cacheOptions.loglevel
+								})
+							).apply(compiler);
+							break;
+						}
+						case "background": {
+							const BackgroundFileCachePlugin = require("./cache/BackgroundFileCachePlugin");
+							const SeparateFilesCacheStrategy = require("./cache/SeparateFilesCacheStrategy");
+							new BackgroundFileCachePlugin(
+								new SeparateFilesCacheStrategy({
+									cacheLocation: cacheOptions.cacheLocation,
+									version: cacheOptions.version,
+									hashAlgorithm: cacheOptions.hashAlgorithm,
+									loglevel: cacheOptions.loglevel
+								})
+							).apply(compiler);
+							break;
+						}
+						case "idle": {
+							const IdleFileCachePlugin = require("./cache/IdleFileCachePlugin");
+							const SeparateFilesCacheStrategy = require("./cache/SeparateFilesCacheStrategy");
+							new IdleFileCachePlugin(
+								new SeparateFilesCacheStrategy({
+									cacheLocation: cacheOptions.cacheLocation,
+									version: cacheOptions.version,
+									hashAlgorithm: cacheOptions.hashAlgorithm,
+									loglevel: cacheOptions.loglevel
+								})
+							).apply(compiler);
+							break;
+						}
+						case "pack": {
+							const IdleFileCachePlugin = require("./cache/IdleFileCachePlugin");
+							const PackFileCacheStrategy = require("./cache/PackFileCacheStrategy");
+							new IdleFileCachePlugin(
+								new PackFileCacheStrategy({
+									cacheLocation: cacheOptions.cacheLocation,
+									version: cacheOptions.version,
+									loglevel: cacheOptions.loglevel
+								})
+							).apply(compiler);
+							break;
+						}
+					}
 					break;
 				}
 				default:
