+++ lib/WebpackOptionsDefaulter.js
@@ -58,6 +58,27 @@ class WebpackOptionsDefaulter extends OptionsDefaulter {
 					value.name =
 						(options.name || "default") + "-" + (options.mode || "production");
 				}
+				if (value.version === undefined) {
+					value.version = "";
+				}
+				if (value.cacheDirectory === undefined) {
+					value.cacheDirectory = "node_modules/.cache/webpack/";
+				}
+				if (value.cacheLocation === undefined) {
+					value.cacheLocation = path.resolve(value.cacheDirectory, value.name);
+				}
+				if (value.hashAlgorithm === undefined) {
+					value.hashAlgorithm = "md4";
+				}
+				if (value.store === undefined) {
+					value.store = "pack";
+				}
+				if (value.idleTimeout === undefined) {
+					value.idleTimeout = 10000;
+				}
+				if (value.idleTimeoutForInitialStore === undefined) {
+					value.idleTimeoutForInitialStore = 0;
+				}
 			}
 			return value;
 		});
